Remove duplicate graphql types
