+++ graphql/src/schema/mutations/inbound_shipment/line/delete.rs
@@ -1,10 +1,11 @@
 use async_graphql::*;
 
-use super::{BatchIsReserved, InvoiceLineBelongsToAnotherInvoice};
+use super::BatchIsReserved;
 use crate::schema::{
     mutations::{
         CannotEditInvoice, DeleteResponse, ForeignKey, ForeignKeyError,
-        InvoiceDoesNotBelongToCurrentStore, NotAnInboundShipment,
+        InvoiceDoesNotBelongToCurrentStore, InvoiceLineBelongsToAnotherInvoice,
+        NotAnInboundShipment,
     },
     types::{DatabaseError, ErrorWrapper, RecordNotFound},
 };
