+++ graphql/src/schema/mutations/inbound_shipment/line/mod.rs
@@ -2,9 +2,6 @@ use async_graphql::*;
 
 pub mod delete;
 
-use crate::schema::queries::invoice::*;
-use crate::ContextExt;
-
 pub use self::delete::*;
 
 pub mod insert;
@@ -18,19 +15,5 @@ pub struct BatchIsReserved;
 impl BatchIsReserved {
     pub async fn description(&self) -> &'static str {
         "Batch is already reserved/issued"
-    }
-}
-
-pub struct InvoiceLineBelongsToAnotherInvoice(String);
-#[Object]
-impl InvoiceLineBelongsToAnotherInvoice {
-    pub async fn description(&self) -> &'static str {
-        "Invoice line belongs to another invoice"
-    }
-
-    pub async fn invoice(&self, ctx: &Context<'_>) -> InvoiceResponse {
-        let connection_manager = ctx.get_connection_manager();
-
-        get_invoice(connection_manager, self.0.clone())
     }
 }
